requirements.txt
streamlit==1.32.0
langchain==0.1.11
transformers==4.38.2
torch==2.2.1
sentence-transformers==2.6.1
faiss-cpu==1.7.4
python-dotenv==1.0.0
wikipedia==1.4.0
duckduckgo-search==3.9.6
accelerate==0.28.0